third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-getStats-promise.html

<!DOCTYPE html>
<html>
<head>
<title>RTCPeerConnection.getStats</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
var pc = new webkitRTCPeerConnection(null);

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({audio:true, video:true})
    .then(function(mediaStream) {
        pc.addStream(mediaStream);
        var selector = pc.getLocalStreams()[0].getVideoTracks()[0];
        assert_not_equals(selector, null);
        return pc.getStats(selector)
          .then(function(statsReport) {
              assert_unreached('The promise should be rejected.');
            }, function(reason) {
              // Expecting promise to be rejected until getStats is implemented.
              // crbug.com/627816
            });
      });
}, 'getStats(MediaStreamTrack selector)');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({audio:true, video:true})
    .then(function(mediaStream) {
        pc.addStream(mediaStream);
        return pc.getStats(null)
          .then(function(statsReport) {
              assert_unreached('The promise should be rejected.');
            }, function(reason) {
              // Expecting promise to be rejected until getStats is implemented.
              // crbug.com/627816
            });
      });
}, 'getStats(null)');

promise_test(function() {
  return navigator.mediaDevices.getUserMedia({audio:true, video:true})
    .then(function(mediaStream) {
        pc.addStream(mediaStream);
        return pc.getStats()
          .then(function(statsReport) {
              assert_unreached('The promise should be rejected.');
            }, function(reason) {
              // Expecting promise to be rejected until getStats is implemented.
              // crbug.com/627816
            });
      });
}, 'getStats()');
</script>
</body>
</html>
